+++ love/src/components/WeatherStation/WeatherStation.container.jsx
@@ -3,7 +3,9 @@ This file is part of LOVE-frontend.
 
 Copyright (c) 2023 Inria Chile.
 
-Developed by Inria Chile.
+Developed by Inria Chile and the Telescope and Site Software team.
+
+Developed for the Vera C. Rubin Observatory Telescope and Site Systems.
 
 This program is free software: you can redistribute it and/or modify it under 
 the terms of the GNU General Public License as published by the Free Software 
@@ -39,12 +41,6 @@ export const schema = {
       description: 'Whether the component has a raw mode version',
       isPrivate: true,
       default: true,
-    },
-    controls: {
-      type: 'boolean',
-      description: "Whether to display controls to configure periods of time'",
-      default: true,
-      isPrivate: false,
     },
     salindex: {
       type: 'number',